+++ packages/react-server/src/metrics/metrics.ts
@@ -3,9 +3,11 @@ import compose from 'koa-compose';
 
 import {KoaNextFunction} from '../types';
 
+const MILLIS_PER_SECOND = 1000;
+const NANOS_PER_MILLIS = 1e6;
 const START_TIME_STATE_KEY = Symbol('startTime');
 
-async function startStatsd(ctx: Context, next: KoaNextFunction) {
+async function startRequestTiming(ctx: Context, next: KoaNextFunction) {
   ctx.state[START_TIME_STATE_KEY] = process.hrtime();
   await next();
 }
@@ -17,11 +19,12 @@ async function middleware(ctx: Context, next: KoaNextFunction) {
     const [seconds, nanoseconds] = process.hrtime(
       ctx.state[START_TIME_STATE_KEY],
     );
-    const ms = seconds * 1000 + nanoseconds / 1e6;
+    const ms = seconds * MILLIS_PER_SECOND + nanoseconds / NANOS_PER_MILLIS;
     const requestTime = Math.round(ms);
 
-    ctx.set('X-React-Server-Request-Time', requestTime.toString());
+    const uiMetrics = `ui;request_time=${requestTime}`;
+    ctx.set('Server-Timing', uiMetrics);
   }
 }
 
-export const metricsMiddleware = compose([startStatsd, middleware]);
+export const metricsMiddleware = compose([startRequestTiming, middleware]);